[LFXV2-1730] refactor: rename and consolidate notification files
- notifications_handler.go → project_subscriber.go
  Rename makes the pattern explicit: event subscribers (fire-and-forget,
  no response) live here; future handlers like ProjectUpdated belong here too
- Fold notifications_diff.go into project_subscriber.go
  diffNewMembers/diffRole/memberKey are only called from the subscriber and
  don't warrant a separate file
- Merge the two test files into project_subscriber_test.go accordingly
- Update log prefixes from "notifications_handler:" to "project_subscriber:"
